+++ src/components/App.js
@@ -1,22 +1,21 @@
 import React from 'react';
 import {BrowserRouter as Router, Route} from 'react-router-dom';
 
-import Navigation from './Navigation';
-import LandingPage from './Landing';
-import SignUpPage from './SignUp';
-import SignInPage from './SignIn';
-import PasswordForgetPage from './PasswordForget';
-import HomePage from './Home';
-import AccountPage from './Account';
 
-import * as routes from '../constants/routes';
-import withAuthentication from './withAuthentication';
+import * as routes from '../shared/routes';
+import withAuthentication from '../shared/withAuthentication';
 
+import LandingPage from '../pages/landing';
+import SignUpPage from '../pages/signup';
+import LogInPage from '../pages/login';
+import ResetPage from '../pages/reset';
+import Dashboard from '../pages/dashboard';
 import Settings from '../pages/settings';
 import Step1 from '../pages/settings/step-1';
 import Step2 from '../pages/settings/step-2';
 import Info from '../pages/info';
-import ProfilePage from '../pages/profile/create';
+
+import Navigation from './Navigation';
 
 const App = () =>
   <Router>
@@ -24,17 +23,15 @@ const App = () =>
       <Navigation/>
       <Route exact path={routes.LANDING} component={() => <LandingPage/>}/>
       <Route exact path={routes.SIGN_UP} component={() => <SignUpPage/>}/>
-      <Route exact path={routes.SIGN_IN} component={() => <SignInPage/>}/>
-      <Route exact path={routes.PASSWORD_FORGET} component={() => <PasswordForgetPage/>}/>
-      <Route exact path={routes.DASHBOARD} component={() => <HomePage/>}/>
-      <Route exact path={routes.PASSWORD_RESET} component={() => <PasswordResetPage/>}/>
+      <Route exact path={routes.LOG_IN} component={() => <LogInPage/>}/>
+      <Route exact path={routes.RESET} component={() => <ResetPage/>}/>
       <Route exact path={routes.DASHBOARD} component={() => <Dashboard/>}/>
-      <Route exact path={routes.ACCOUNT} component={() => <AccountPage/>}/>
       <Route exact path={routes.SETTINGS} component={() => <Settings/>}/>
       <Route exact path={routes.STEP_1} component={() => <Step1/>}/>
       <Route exact path={routes.STEP_2} component={() => <Step2/>}/>
       <Route exact path={routes.INFO} component={() => <Info/>}/>
     </div>
-  </Router>;
+  </Router>
+;
 
 export default withAuthentication(App);
