+++ src/components/Forms.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import {reduxForm} from 'redux-form';
 
+
 let GroupsForm = props => {
   return (
     <form className={props.className} onSubmit={props.onSubmit}>{props.children}</form>
@@ -10,6 +11,7 @@ let GroupsForm = props => {
 GroupsForm = reduxForm({
   form: 'userGroups'
 })(GroupsForm);
+
 
 let ProductsForm = props => {
   return (
