--- src/components/SignOut.js
@@ -1,6 +1,0 @@
-
-const SignOutButton = () =>
-  <button type="button"
-          onClick={auth.doSignOut}>Sign Out</button>;
-
-export default SignOutButton;
