--- src/components/helpers.js
@@ -1,5 +1,0 @@
-export const updateByPropertyName = (propertyName, value) => () => ({
-    [propertyName]: value,
-});
-
-export const authCondition = (authUser) => !!authUser;
