+++ src/firebase/db.js
@@ -1,16 +1,13 @@
-import { db } from './firebase';
-import * as firebase from 'firebase';
-
-const db = firebase.database();
+import {db} from './config';
 
 export const getProducts = () =>
   db.ref(`products`).once('value');
 
 export const createUser = (id, username, email) =>
-    db.ref(`users/${id}`).set({
-        username,
-        email
-    });
+  db.ref(`users/${id}`).set({
+    username,
+    email
+  });
 
 export const onceGetUsers = () =>
-    db.ref('users').once('value');
+  db.ref('users').once('value');
