--- src/firebase/index.js
@@ -1,7 +1,0 @@
-import * as auth from './auth';
-import * as db from './db';
-import * as firebase from './firebase';
-
-export {
-  auth, firebase, db
-};
