--- src/helpers.js
@@ -1,3 +1,0 @@
-export const propByKey = (propertyName, value) => () => ({
-  [propertyName]: value,
-});
