--- src/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import sessionReducer from './session';
-import userReducer from './user';
-
-const rootReducer = combineReducers({
-    sessionState: sessionReducer,
-    userState: userReducer,
-});
-
-export default rootReducer;
