+++ src/shared/helpers.js
@@ -1,0 +1,5 @@
+export const updateByPropertyName = (propertyName, value) => () => ({
+    [propertyName]: value,
+});
+
+export const authCondition = (authUser) => !!authUser;
