+++ src/shared/routes.js
@@ -1,12 +1,10 @@
-export const HOME = '/home';
-export const ACCOUNT = '/account';
-export const PASSWORD_FORGET = '/pw_forget';
 export const LANDING = '/';
 export const SIGN_UP = '/rejestracja';
-export const SIGN_IN = '/logowanie';
-export const INFO = '/informacje';
+export const LOG_IN = '/logowanie';
+export const RESET = '/resetuj';
 export const DASHBOARD = '/pulpit';
 export const SETTINGS = '/ustawienia';
 export const STEP_1 = '/ustawienia/krok-1';
 export const STEP_2 = '/ustawienia/krok-2';
+export const INFO = '/informacje';
