+++ src/state/store.js
@@ -1,6 +1,6 @@
 import {createStore, applyMiddleware} from 'redux';
 import thunkMiddleware from 'redux-thunk';
-import rootReducer from './reducers';
+import rootReducer from './reducers/index';
 
 const store = createStore(
   rootReducer,
