--- src/store/index.js
@@ -1,6 +1,0 @@
-import { createStore } from 'redux';
-import rootReducer from '../reducers';
-
-const store = createStore(rootReducer);
-
-export default store;
